WeightToFee
