+++ runtime/common/Cargo.toml
@@ -11,7 +11,8 @@ version     = "6.0.0"
 
 [dependencies]
 # crates.io
-codec = { package = "parity-scale-codec", version = "3.2", default-features = false }
+codec    = { package = "parity-scale-codec", version = "3.2", default-features = false }
+smallvec = { version = "1.10" }
 
 # darwinia
 dc-primitives = { default-features = false, path = "../../core/primitives" }
