+++ runtime/common/src/lib.rs
@@ -27,6 +27,17 @@ pub use bp_darwinia_core as bp_crab;
 pub use bp_darwinia_core as bp_darwinia;
 pub use bp_darwinia_core as bp_pangolin;
 
+// darwinia
+use dc_primitives::*;
+// substrate
+use frame_support::{
+	sp_runtime::Perbill,
+	weights::{
+		constants::ExtrinsicBaseWeight, WeightToFeeCoefficient, WeightToFeeCoefficients,
+		WeightToFeePolynomial,
+	},
+};
+
 #[macro_export]
 macro_rules! fast_runtime_or_not {
 	($name:ident, $development_type:ty, $production_type:ty) => {
@@ -35,4 +46,32 @@ macro_rules! fast_runtime_or_not {
 		#[cfg(not(feature = "fast-runtime"))]
 		type $name = $production_type;
 	};
+}
+
+/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
+/// node's balance type.
+///
+/// This should typically create a mapping between the following ranges:
+///   - `[0, MAXIMUM_BLOCK_WEIGHT]`
+///   - `[Balance::min, Balance::max]`
+///
+/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
+///   - Setting it to `0` will essentially disable the weight fee.
+///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
+pub struct WeightToFee;
+impl WeightToFeePolynomial for WeightToFee {
+	type Balance = Balance;
+
+	fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+		// in Rococo, extrinsic base weight (smallest non-zero weight) is mapped to 1 MILLIUNIT:
+		// here, we map to 1/10 of that, or 1/10 MILLIUNIT
+		let p = MILLIUNIT / 10;
+		let q = 100 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
+		smallvec::smallvec![WeightToFeeCoefficient {
+			degree: 1,
+			negative: false,
+			coeff_frac: Perbill::from_rational(p % q, q),
+			coeff_integer: p / q,
+		}]
+	}
 }
